requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.8
dj-database-url==0.5.0
Django==2.2.6
django-anymail==8.4
django-heroku==0.3.1
gunicorn==20.1.0
idna==3.3
Pillow==8.4.0
psycopg2==2.8.6
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
urllib3==1.26.7
whitenoise==5.3.0
